@@ -5,11 +5,10 @@ author: Tony Borchert
 homepage:
 
 dependencies:
-#  flutter:
-#    sdk: flutter
   meta: any
+
 dev_dependencies:
-  test: ^1.15.7
+  test: ^1.16.8
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.2 <3.0.0"
