@@ -2,25 +2,29 @@ import 'package:meta/meta.dart';
 import 'constants.dart';
 
 @immutable
-abstract class DebugObject {
+abstract class DebugObject {}
+
+
+@immutable
+abstract class TextDebugObject {
   final int debugLine;
   final int debugCharacter;
-  final int secondCharacter;
+  final int? secondCharacter;
 
-  DebugObject(this.debugLine, this.debugCharacter, [this.secondCharacter]);
+  TextDebugObject(this.debugLine, this.debugCharacter, [this.secondCharacter]);
 
   @override
   bool operator ==(Object other) =>
-      other is DebugObject &&
+      other is TextDebugObject &&
       other.debugLine == this.debugLine &&
       other.debugCharacter == this.debugCharacter &&
       other.secondCharacter == this.debugCharacter;
 }
 
-@immutable
-abstract class TSException {
+class TSException {
+  TSException._();
   static String generateErrorShow(String line, int character,
-          [int secondDebugCharacter]) =>
+          [int? secondDebugCharacter]) =>
       error_space +
       line +
       "\n" +
@@ -28,12 +32,4 @@ abstract class TSException {
       (" " * character) +
       ("^" * (((secondDebugCharacter ?? character) + 1) - character)) +
       "\n";
-
-  final String message;
-  final int debugLine;
-  final int debugCharacter;
-  final int secondDebugCharacter;
-
-  TSException(this.message, this.debugLine, this.debugCharacter,
-      this.secondDebugCharacter);
 }
