@@ -1,49 +1,50 @@
 import 'package:test/test.dart';
 import 'package:tsharp/future_values/future_values.dart';
-import 'package:tsharp/direct_values/values.dart';
+import 'package:tsharp/direct_values/direct_values.dart';
 import 'package:tsharp/parsing/parse_error_handling.dart';
 
 import 'package:tsharp/parsing/value_parsing.dart';
 import 'package:tsharp/parsing/parse_debug.dart';
 import 'package:tsharp/parsing/token_parsing.dart';
-
+import 'package:tsharp/parsing/extensions.dart';
 import 'package:tsharp/debug.dart';
 
 class O extends Operator {
-  O(String operator) : super(operator, null, null, null);
+  O(String operator) : super(operator, 0, 0, 0);
 }
+
+ParseDebugStream ns = ParseDebugStream();
 
 void main() {
   group("Simple value parsing", () {
     test("Value parsing", () {
       expect(
-        (parseValue("_", 1, 1, null) as PrimitiveValue).value,
+        (parseValue("_", 1, 1, ns) as PrimitiveValue).value,
         SpecialValues.absent,
       );
-      expect((parseValue("absent", 1, 1, null) as PrimitiveValue).value,
+      expect((parseValue("absent", 1, 1, ns) as PrimitiveValue).value,
           SpecialValues.absent);
-      expect((parseValue("23", 1, 1, null) as PrimitiveValue).value, 23);
-      expect((parseValue("23.0", 1, 1, null) as PrimitiveValue).value, 23.0);
-      expect((parseValue("\"asdf{\\\\\"", 1, 1, null) as PrimitiveValue).value,
+      expect((parseValue("23", 1, 1, ns) as PrimitiveValue).value, 23);
+      expect((parseValue("23.0", 1, 1, ns) as PrimitiveValue).value, 23.0);
+      expect((parseValue("\"asdf{\\\\\"", 1, 1, ns) as PrimitiveValue).value,
           "asdf{\\");
       expect(
-          ((parseValue("\$:5", 1, 1, null) as OperatorCall).parameters.first
+          ((parseValue("\$:5", 1, 1, ns) as OperatorCall).parameters.first
                   as RecordReference)
               .invocation,
           "params");
     });
     test("Bad value parsing", () {
-      const val = "\"\"dkks";
+      const val = "=D";
       //anderes beispiel das nicht funktionier = "asdf (asd, asd)"
       List<String> split = val.split("\n");
       final stream = ParseDebugStream();
       try {
         final result = parseValue(val, 1, 1, stream);
-        print(result);
       } catch (error) {
         stream.processException(error);
       }
-      print(stream.asErrorLog("[TESTING]", split));
+      // print(stream.asErrorLog("[TESTING]", split));
     });
   });
 
@@ -52,15 +53,15 @@ void main() {
       expect(
           Operator.leastImportant([O("&&"), O("*"), O("-"), O(":")]).operator,
           "&&");
-      expect(Operator.leastImportant([O("-"), O("*"), O("....")]).operator,
-          "-");
+      expect(
+          Operator.leastImportant([O("-"), O("*"), O("....")]).operator, "-");
 
-      expect((parseValue("234+234 * 4", 1, 1, null) as OperatorCall).function,
-          "+");
+      expect(
+          (parseValue("234+234 * 4", 1, 1, ns) as OperatorCall).function, "+");
     });
 
     test("Prefixes and postfixes", () {
-      final parsePrefix = parseValue("!asdf..!", 1, 1, null) as PrefixCall;
+      final parsePrefix = parseValue("!asdf..!", 1, 1, ns) as PrefixCall;
       expect(parsePrefix.function, "!");
       expect(parsePrefix.debugLine, 1);
       expect(parsePrefix.debugCharacter, 1);
@@ -89,7 +90,7 @@ void main() {
     test("Token parsing", () {
       const val = "if ads == ksdf{\"{{{{{\"}else{}\n"
           "       return(bubbel) + one\n{}()";
-      expect(parseToTokens(val, 1, 1, null), [
+      expect(parseToTokens(val, 1, 1, ns), [
         [
           Token("if", 1, 1),
           Token("ads == ksdf", 1, 4, false),
@@ -99,10 +100,10 @@ void main() {
         ],
         [
           Token("return", 2, 8),
-          Token("(bubbel) + one", 2, 14, false)
+          Token("(bubbel) + one", 2, 14, false),
         ],
         [
-          Token("{}()", 3, 1)
+          Token("{}()", 3, 1),
         ],
       ]);
     });
@@ -110,13 +111,14 @@ void main() {
 
   test("Function call parsing", () {
     final parseValueResult = parseValue(
-        "add(     [\n"
-        "3423,\n"
-        "         \"je susis{{{{{{\\\\\\\"{\",\n"
-        "     23456787654],           \"\\\"\\\\\")",
-        1,
-        1,
-        null) as FunctionCall;
+      "add(     [\n"
+      "3423,\n"
+      "         \"je susis{{{{{{\\\\\\\"{\",\n"
+      "     23456787654],           \"\\\"\\\\\")",
+      1,
+      1,
+      ns,
+    ) as FunctionCall;
     expect(parseValueResult.runtimeType, FunctionCall);
 
     final parseValueResultFunction =
@@ -158,13 +160,19 @@ void main() {
 
   test("instruction grouping", () {
     final parseTestString = """
-        var    adfgfd    **=   {
+        var    adfgfd    **=  else{\n}\nkdjs""";
 
+    final List<List<Token>> parseResult =
+        parseToTokens(parseTestString, 1, 1, ns);
+    final List<Token> tokens = [];
+    for (List<Token> tokenList in parseResult) tokens.addAll(tokenList);
+    final combineResult = tokens.combine();
+    // print(combineResult.token);
+  });
 
-
-}bdfdfdfdf     +d     csdfsdf
-""";
-
-    final parseResult = parseToTokens(parseTestString, 1, 1, null);
+  group("Test the stream that should not be used", () {
+    test("", () {
+      expect(ns.events.length, 0);
+    });
   });
 }
